requirements.txt
streamlit
google-generativeai
python-dotenv
psycopg2-binary
cryptography
openai
pandas
plotly
matplotlib